ai-advisor-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
pandas==2.1.4
numpy==1.26.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

